[project @ 2002-01-14 18:23:10 by edburns] The problem was a lifetime bug with RenderContext.  The new ObjectAccessor favors things being in request scope (due to its first param being hard coded to ServletRequest) That said, many of the ModelAccessor callsites now call objectAccessor.getObject(renderContext.getRequest(), name), but the renderContext was being placed in session scope, and thus its request was not valid after the initial request that caused the renderContext to get created.
I have a call in to Amy to find out what is the correct lifetime of a
RenderContext instance.  I think it is request, in which case we'll have
to rethink our usage of it throughout the code.

In the mean time, my fix is to add an ri specific method to our
RenderContext implementation that allows its request to be set.

Original author: edburns
Date: 2002-01-14 18:23:10+00:00

git-svn-id: https://svn.java.net/svn/mojarra~svn/trunk@170 761efcf2-d34d-6b61-9145-99dcacc3edf1