[project @ 2002-07-24 01:13:27 by craigmcc] Make UIComponent extend Serializable, so that component implementation classes will be Serializable.
NOTE:  For testing purposes, this is the only class I've changed.  If we
do it for real, we'd probably want to make Tree serializable as well (so
that you could just call getResponseTree() and start there), but Tree
contains a reference to RenderKit -- which is not generally going to be
something you want to serialize.  So, for now, plan on having to
serialize starting at getResponseTree().getRoot(), starting at
getResponseTree().getTreeId(), and separately save the render kit identifier
so that it can all be restored.  (The latter might be problematic because
you can't ask a RenderKit instance for its id -- for now I'm ok with punting
on this and just assuming it is the default id).

Original author: craigmcc
Date: 2002-07-24 01:13:27+00:00

git-svn-id: https://svn.java.net/svn/mojarra~svn/trunk@540 761efcf2-d34d-6b61-9145-99dcacc3edf1