[project @ 2003-08-14 23:14:43 by horwat] [r=edburns]
This bug fix addresses the TestRenderResponsePhase error that occurred on Solaris and Linux. Turns out that Jasper was throwing an java.lang.ArrayIndexOutOfBoundsException: 3584 when attempting to write the closing bracket char. This is a Jasper bug in the char writer. Ryan discovered that once the buffer size limit is reached the new buffer is reallocated with an incorrect length count of 0.

This has been fixed in the Jasper nightly but would be a problem in JWSDP 1.2 container.

Original author: horwat
Date: 2003-08-14 23:14:43+00:00

git-svn-id: https://svn.java.net/svn/mojarra~svn/trunk@1655 761efcf2-d34d-6b61-9145-99dcacc3edf1