[project @ 2003-09-09 21:15:26 by craigmcc] Clarify the DTD such that <summary> and <detail> elements with no "xml:lang" should be associated with the language associated wth the default Locale for the underlying JVM.
Clarify MessageResources() such that, when localized text for a
particular Locale is not available, it can return the text associated
with the language specified by the default Locale (i.e. by the text
that was configured with no xml:lang attributes).

RI IMPACT:  Verify that current behavior matches this (I think it does).

Original author: craigmcc
Date: 2003-09-09 21:15:27+00:00

git-svn-id: https://svn.java.net/svn/mojarra~svn/trunk@1844 761efcf2-d34d-6b61-9145-99dcacc3edf1