[project @ 2003-10-21 05:37:43 by craigmcc] Beginnings of remodel for state saving and restoring in Repeater components (i.e. UIData).  The end objective is to allow child components to implement their ValueHolder et. al. properties normally, without having to be aware that they are inside a Repeater.
All of the unit tests currently work, but UIData is not currently capable
of saving and restoring the per-row state of its children in a complete
manner.  I wanted to localize the impact of my remaining changes to avoid
conflicts with other checkins on the HEAD of jsf-api.

PENDING:  these changes make the value of having ValueHolderSupport and
ConvertableValueHolderSupport minimal again, so I will probably remove them
if nothing else shows up to motivate their retention.

PENDING:  Unit tests for UIData that exercise the per-row state saving
and restoring capabilities.

PENDING:  Factor the per-row stuff out into a new-style RepeaterSupport
if and only if it makes sense.

Original author: craigmcc
Date: 2003-10-21 05:37:49+00:00

git-svn-id: https://svn.java.net/svn/mojarra~svn/trunk@2178 761efcf2-d34d-6b61-9145-99dcacc3edf1