[project @ 2003-11-07 18:55:28 by craigmcc] Scarab Issue:  [JSFA80] Value Bindings - Phase II
The previous algorithm of erasing a value binding when you explicitly set
the property value is not what we want for value (when we remove valueRef)
because we need the valueRef to stay around when the page is redisplayed.
Therefore, change the rules such that an explicitly set attribute or property
"shadows" any value binding for the same attribute or property.  You can
unshadown an attribute or non-primitive property by setting the value to null.
Unfortunately, there's no current way to unshadow a primitive property.

RI IMPACT:  No differences from the impacts in the Phase I check
  (but the Phase III checkin will definitely have impact)

TCK IMPACT:  Don't even bother looking at the value bindings stuff until
it settles down a little more.

DEMO IMPACT:  We'll need to remodel our demo components and modify the tags
so that they support deferred evaluation as well.  But that can wait until
the APIs settle down.

Original author: craigmcc
Date: 2003-11-07 18:55:41+00:00

git-svn-id: https://svn.java.net/svn/mojarra~svn/trunk@2330 761efcf2-d34d-6b61-9145-99dcacc3edf1