[project @ 2004-01-12 06:52:58 by craigmcc] [JSFA148]
With the latest changes to add the "submittedValue" property, an application
action now needs to behave differenty if it is invoked in Apply Request
Values (immediate=true) versus Invoke Application (immediate=false).  In the
former case, the "value" property has not been updated, so the "checked"
attribute has to examine the "submittedValue" property -- and the fact
that CheckboxRenderer sets this to "true" or "false" is an implementation
detail, not something that is specified.  But at least it works in both
modes now, with the JSFA148 patch that Hans suggested.

Original author: craigmcc
Date: 2004-01-12 06:52:59+00:00

git-svn-id: https://svn.java.net/svn/mojarra~svn/trunk@2477 761efcf2-d34d-6b61-9145-99dcacc3edf1