Added RELEASE_PENDING to getCurentComponent().
Need to re-evaluate this method.  The original intent was calling this from the EL, however, we have the component implicit object, so exposing this in an EL friendly way isn't really necessary.  It may be useful to keep the method for developers to leverage in their code, but it should accept a FacesContext to avoid additional ThreadLocal lookups.

git-svn-id: https://svn.java.net/svn/mojarra~svn/trunk@5076 761efcf2-d34d-6b61-9145-99dcacc3edf1