Change renderkit to allow for user wrapped onchange and onclick events with ajax calls.
Include full tests.

Two issues remain - commandButton onchange is now listed (possibly incorrectly as passthrough false) - I've modified Attribute manager to make it behave as passthrough, but we may wish to modify this.

Also, in the event that a commandButton has both an ajax tag and a param tag, the param tag is ignored.  I'm awaiting the EG's ruling on this matter.

git-svn-id: https://svn.java.net/svn/mojarra~svn/branches/AJAX_PFD_BRANCH@5919 761efcf2-d34d-6b61-9145-99dcacc3edf1