Issue: 451 part 2
SECTION: Modified Files
----------------------------
M      jsf-api/doc/standard-html-renderkit-base.xml
M      jsf-api/doc/selectmany-props.xml
M      jsf-api/doc/standard-html-renderkit.xml
M      jsf-api/doc/selectone-radio-props.xml

- Spec changes for hideNoSelectionLabel.  These are

* add hideNoSelectionAttribute to select* components

Flag indicating that, if this component is activated by the user, The
"no selection option", if any, must be hidden.

* add this text to "rendering the option elements"

If the current child is a SelectItem whose noSelctionProperty is true,
and the UISelectOne or UISelectMany parent of this option has one or
more selected values that are not the "no selection" SelectItem, and the
component has a "hideNoSelectionLabel" attribute whose value is true,
then the current option, which is the "no selection" option, must not be
rendered.


git-svn-id: https://svn.java.net/svn/mojarra~svn/trunk@6026 761efcf2-d34d-6b61-9145-99dcacc3edf1