Issue 1037 - Merge javax.faces.partial.source and javax.faces.behavior.source request params
These two request parameters have been merged into the "javax.faces.source" parameter.

This commit also has a minor id handling fix.  Previously, if no id is specified on the component, we do not write an id to the client.  This causes problems for <f:ajax>, which needs access to the client id in order to populate the javax.faces.source request parameter.  To address this, HtmlBasicRenderer.shouldWriteIdAttribute() now returns true when client components are present.

r=rogerk

git-svn-id: https://svn.java.net/svn/mojarra~svn/trunk@6815 761efcf2-d34d-6b61-9145-99dcacc3edf1