Fix for bug 1096.
Thanks to Alexi Smirnov for the suggested patch, and the bug report.

Ajax requests were causing a new View to be computed each time, which rolled the previous views off of the cache.  After a reletively small number of ajax requests, using the back button would trigger a ViewExpiredException.

This patch fixes that, but may require a spec change - see spec bug 549 for more details.



git-svn-id: https://svn.java.net/svn/mojarra~svn/trunk@7230 761efcf2-d34d-6b61-9145-99dcacc3edf1