Initial checkin for JSF Forum demo.
Just a prototype, and lots to do, but it works - though it's not concurrency-safe (it uses file read and write without even trying to address locking - we'll eventually move this to a database).

The goal of this demo - end to end demonstration of JSF 2 best practices, which creating a theoretically useful peice of software.

git-svn-id: https://svn.java.net/svn/mojarra~svn/trunk@7533 761efcf2-d34d-6b61-9145-99dcacc3edf1